fix: decouple results from viewport, make region-driven (#93)
## What changed

### Removed
- **useViewportResorts.js** — deleted entirely. No more viewport bounds
  queries on every moveend firing across 4K resorts.
- **renderedResorts / filteredResorts** from Zustand store — these were
  viewport-derived state that caused flicker during flyTo animations.

### Added/Changed
- **Region-driven displayedResorts** in page.js — when user selects a
  region, results = all resorts in that region filtered by pass toggles.
  No viewport dependency. Zero flicker. Sorted by 7-day snowfall desc,
  capped at 100 results.
- **currentZoom** still updated via simple setCurrentZoom in MapExplore's
  onMoveEnd handler (needed for satellite toggle, snow layers, etc).
- Map markers remain viewport-aware via Mapbox layer minzoom/maxzoom.
  Only the RESULTS LIST is now region-driven.

### Why
Every moveend was firing a bounds query on 4K resorts → writing to
Zustand → components re-rendering. During flyTo animations this caused
flicker (stale → empty → new data). The architecture was fundamentally
wrong — results should be driven by the user's selected region, not
the transient viewport bounds.
