+++ src/app/page.js
@@ -25,13 +25,13 @@ function AppContent() {
 
   const selectedResort = useMapStore((s) => s.selectedResort);
   const setSelectedResort = useMapStore((s) => s.setSelectedResort);
-  const filteredResorts = useMapStore((s) => s.filteredResorts);
   const searchQuery = useMapStore((s) => s.searchQuery);
   const showIkon = useMapStore((s) => s.showIkon);
   const showEpic = useMapStore((s) => s.showEpic);
   const showMC = useMapStore((s) => s.showMC);
   const showIndy = useMapStore((s) => s.showIndy);
   const showIndependent = useMapStore((s) => s.showIndependent);
+  const snowBySlug = useMapStore((s) => s.snowBySlug);
 
   // Active passes set — drives carousel/sidebar filtering
   const activePasses = useMemo(() => {
@@ -44,18 +44,20 @@ function AppContent() {
     return passes;
   }, [showIkon, showEpic, showMC, showIndy, showIndependent]);
 
-  // Dynamic results: search overrides viewport, otherwise show viewport resorts
+  // Region-driven results: no viewport dependency, zero flicker.
+  // When user selects a region, results = all resorts in that region filtered by pass toggles.
   const displayedResorts = useMemo(() => {
     const query = (searchQuery || "").toLowerCase().trim();
 
     if (query) {
+      // Search across ALL resorts
       return resorts.filter((r) => {
         const p = r.properties;
         if (!activePasses.has(p?.pass)) return false;
         const name = (p?.name || "").toLowerCase();
         const state = (p?.state || "").toLowerCase();
         const country = (p?.country || "").toLowerCase();
-        const region = (p?.region || "").toLowerCase();
+        const region = (p?.region_id || "").toLowerCase();
         return name.includes(query) || state.includes(query) || country.includes(query) || region.includes(query);
       });
     }
@@ -63,17 +65,30 @@ function AppContent() {
     // At globe view, show nothing — user picks a region first
     if (nav.isGlobe) return [];
 
-    // Filter viewport resorts by active passes + current region
-    // Region filter prevents stale global results from showing during flyTo
-    // (viewport bounds at globe zoom contain the entire world)
-    return filteredResorts.filter((r) => {
-      if (!activePasses.has(r.properties?.pass)) return false;
-      if (nav.region) {
-        return r.properties?.region_id === nav.region;
-      }
+    // Region-driven: filter by region + active passes
+    const regionId = nav.region;
+    let results = resorts.filter((r) => {
+      const p = r.properties;
+      if (!activePasses.has(p?.pass)) return false;
+      const resortRegion = p?.region_id;
+      if (regionId) return resortRegion === regionId;
       return true;
     });
-  }, [filteredResorts, resorts, activePasses, searchQuery, nav.isGlobe, nav.region]);
+
+    // Sort by 7-day snowfall descending
+    results.sort((a, b) => {
+      const snowA = snowBySlug[a.properties?.slug];
+      const snowB = snowBySlug[b.properties?.slug];
+      const s7dA = snowA?.snowfall_7d || 0;
+      const s7dB = snowB?.snowfall_7d || 0;
+      if (s7dB !== s7dA) return s7dB - s7dA;
+      const avgA = parseFloat(a.properties?.avg_snowfall) || 0;
+      const avgB = parseFloat(b.properties?.avg_snowfall) || 0;
+      return avgB - avgA;
+    });
+
+    return results.slice(0, 100); // Cap at 100
+  }, [resorts, activePasses, searchQuery, nav.isGlobe, nav.region, snowBySlug]);
 
   return (
     <div className="flex h-[calc(100dvh-3rem)] flex-col overflow-hidden sm:h-[calc(100dvh-3.5rem)]">
